+++ Algorithms/pytorch/PPO/ppo.py
@@ -15,7 +15,7 @@ from Common.GAE import estimate_advantages
 from Common.MemoryCollector import MemoryCollector
 from Utils.env_util import get_env_info
 from Utils.file_util import check_path
-from Utils.torch_util import DOUBLE, device
+from Utils.torch_util import device, FLOAT
 from Utils.zfilter import ZFilter
 
 
@@ -60,11 +60,11 @@ class PPO:
         self.env.seed(self.seed)
 
         if env_continuous:
-            self.policy_net = Policy(num_states, num_actions).double().to(device)
+            self.policy_net = Policy(num_states, num_actions).to(device)
         else:
-            self.policy_net = DiscretePolicy(num_states, num_actions).double().to(device)
+            self.policy_net = DiscretePolicy(num_states, num_actions).to(device)
 
-        self.value_net = Value(num_states).double().to(device)
+        self.value_net = Value(num_states).to(device)
         self.running_state = ZFilter((num_states,), clip=5)
 
         if self.model_path:
@@ -81,7 +81,7 @@ class PPO:
 
     def choose_action(self, state):
         """select action"""
-        state = DOUBLE(state).unsqueeze(0).to(device)
+        state = FLOAT(state).unsqueeze(0).to(device)
         with torch.no_grad():
             action, log_prob = self.policy_net.get_action_log_prob(state)
         return action, log_prob
@@ -122,12 +122,12 @@ class PPO:
                             }, i_iter)
 
         batch = memory.sample()  # sample all items in memory
-
-        batch_state = DOUBLE(batch.state).to(device)
-        batch_action = DOUBLE(batch.action).to(device)
-        batch_reward = DOUBLE(batch.reward).to(device)
-        batch_mask = DOUBLE(batch.mask).to(device)
-        batch_log_prob = DOUBLE(batch.log_prob).to(device)
+        #  ('state', 'action', 'reward', 'next_state', 'mask', 'log_prob')
+        batch_state = FLOAT(batch.state).to(device)
+        batch_action = FLOAT(batch.action).to(device)
+        batch_reward = FLOAT(batch.reward).to(device)
+        batch_mask = FLOAT(batch.mask).to(device)
+        batch_log_prob = FLOAT(batch.log_prob).to(device)
 
         with torch.no_grad():
             batch_value = self.value_net(batch_state)
