+++ Algorithms/pytorch/REINFORCE/reinforce.py
@@ -12,7 +12,7 @@ from Algorithms.pytorch.REINFORCE.reinforce_step import reinforce_step
 from Common.MemoryCollector import MemoryCollector
 from Utils.env_util import get_env_info
 from Utils.file_util import check_path
-from Utils.torch_util import DOUBLE, device
+from Utils.torch_util import FLOAT, device
 from Utils.zfilter import ZFilter
 
 
@@ -49,9 +49,9 @@ class REINFORCE:
         self.env.seed(self.seed)
 
         if env_continuous:
-            self.policy_net = Policy(num_states, num_actions).double().to(device)  # current policy
+            self.policy_net = Policy(num_states, num_actions).to(device)  # current policy
         else:
-            self.policy_net = DiscretePolicy(num_states, num_actions).double().to(device)
+            self.policy_net = DiscretePolicy(num_states, num_actions).to(device)
 
         self.running_state = ZFilter((num_states,), clip=5)
 
@@ -68,7 +68,7 @@ class REINFORCE:
 
     def choose_action(self, state):
         """select action"""
-        state = DOUBLE(state).unsqueeze(0).to(device)
+        state = FLOAT(state).unsqueeze(0).to(device)
         with torch.no_grad():
             action, log_prob = self.policy_net.get_action_log_prob(state)
         return action, log_prob
@@ -111,10 +111,10 @@ class REINFORCE:
 
         batch = memory.sample()  # sample all items in memory
 
-        batch_state = DOUBLE(batch.state).to(device)
-        batch_action = DOUBLE(batch.action).to(device)
-        batch_reward = DOUBLE(batch.reward).to(device)
-        batch_mask = DOUBLE(batch.mask).to(device)
+        batch_state = FLOAT(batch.state).to(device)
+        batch_action = FLOAT(batch.action).to(device)
+        batch_reward = FLOAT(batch.reward).to(device)
+        batch_mask = FLOAT(batch.mask).to(device)
 
         p_loss = torch.empty(1)
         for _ in range(self.reinforce_epochs):
